Update imports to use type instead of import
