+++ src/app/modules/game/components/card-list/card-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
-import { Card } from '../../models/card';
+import type { Card } from '../../types';
 
 @Component({
   selector: 'app-card-list',
