+++ src/app/modules/game/components/top/top.component.ts
@@ -1,13 +1,7 @@
 import { Component, EventEmitter, OnInit, Output } from '@angular/core';
 import { Router } from '@angular/router';
 import { GAME_DIFFICULTY } from '../../constants/game-difficulty';
-
-type Difficulty = {
-  label: string;
-  num: number;
-  level: number;
-  icon: string;
-};
+import type { Difficulty } from '../../types';
 
 @Component({
   selector: 'app-top',
