--- src/app/modules/game/enums/game-status.enum.ts
@@ -1,7 +1,0 @@
-/**
- * Game status.
- * - 'NotPlaying': Indicate that a user is not playing the game
- * - 'Playing': Indicate that a user is playing the game
- * - 'Clear': Indicate that a user has flipped all the cards correctly
- */
-export type GameStatus = 'NotPlaying' | 'Playing' | 'Clear';
