--- src/app/modules/game/models/card.ts
@@ -1,10 +1,0 @@
-export type Card = {
-  /** ID */
-  id: number;
-  /** Describes if the card is flipped */
-  flipped: boolean;
-  /** Gets card character.  */
-  character: string;
-  /** Whether the card is done.  */
-  done: boolean;
-};
